feat: Migrate to lazy.nvim as package manager, missing LSP
