--- nvim/after/plugin/airline.lua
@@ -1,23 +1,0 @@
-vim.g['airline#extensions#tabline#enabled'] = 1
-vim.g['airline#extensions#tabline#left_sep'] = ''
-vim.g['airline#extensions#tabline#left_alt_sep'] = ''
-vim.g['airline#extensions#tabline#right_sep'] = ''
-vim.g['airline#extensions#tabline#right_alt_sep'] = ''
-
-vim.g['airline_powerline_fonts'] = 1
-vim.g['airline_left_sep'] = ''
-vim.g['airline_left_alt_sep'] = ''
-vim.g['airline_right_sep'] = ''
-vim.g['airline_right_alt_sep'] = ''
-vim.g['airline_symbols.branch'] = ''
-vim.g['airline_symbols.colnr'] = ' ℅:'
-vim.g['airline_symbols.readonly'] = ''
-vim.g['airline_symbols.linenr'] = ' :'
-vim.g['airline_symbols.maxlinenr'] = '☰ '
-vim.g['airline_symbols.dirty'] = '⚡'
-
-vim.g['airline_theme'] = 'base16_gruvbox_dark_pale'
-
--- Settings
-vim.opt.showtabline = 2
-vim.opt.showmode = false
