--- nvim/after/plugin/chatgpt.lua
@@ -1,3 +1,0 @@
--- require("chatgpt").setup({
--- 
--- })
