--- nvim/after/plugin/markdownpreview.lua
@@ -1,3 +1,0 @@
--- Remaps for markdown preview
-
-vim.keymap.set("n", "<leader>mp", vim.cmd.MarkdownPreview)
