--- nvim/after/plugin/nerdtree.lua
@@ -1,2 +1,0 @@
-vim.keymap.set('n', '<leader>o', vim.cmd.NERDTreeToggle)
-vim.g.NerdTreeShowHidden = true
