--- nvim/after/plugin/venv-selector.lua
@@ -1,4 +1,0 @@
-require("venv-selector").setup()
-
-vim.keymap.set("n", "<leader>vs", ":VenvSelect<CR>")
-vim.keymap.set("n", "<leader>vc", ":VenvSelectCached<CR>")
