+++ nvim/lua/dmacario/init.lua
@@ -1,7 +1,9 @@
 print("Config sourced!")
+
 require("dmacario.remap")
 require("dmacario.set")
+require("dmacario.lazy_init")
 
 vim.o.background = "dark" -- or "light" for light mode
-
+-- vim.cmd([[colorscheme gruvbox]])
