+++ nvim/lua/dmacario/lazy/indent-blankline.lua
@@ -1,0 +1,6 @@
+return{
+    'lukas-reineke/indent-blankline.nvim',
+    config = function()
+        require("ibl").setup()
+    end
+}
