+++ nvim/lua/dmacario/lazy/init.lua
@@ -1,0 +1,32 @@
+return{
+    {
+        'nvim-lua/plenary.nvim',
+        name = "plenary"
+    },
+    'vim-airline/vim-airline-themes',
+    'ryanoasis/vim-devicons',
+    'lewis6991/gitsigns.nvim',
+    'nvim-tree/nvim-web-devicons',
+    'romgrk/barbar.nvim',
+    'stevearc/dressing.nvim',
+    'airblade/vim-gitgutter',
+    'lervag/vimtex',  -- Latex filetype plugin
+    'Vimjas/vim-python-pep8-indent',
+    {
+        'VonHeikemen/lsp-zero.nvim',
+        branch = 'v3.x',
+        dependencies = {
+            --- Uncomment these if you want to manage LSP servers from neovim
+            {'williamboman/mason.nvim'},
+            {'williamboman/mason-lspconfig.nvim'},
+
+            -- LSP Support
+            {'neovim/nvim-lspconfig'},
+            -- Autocompletion
+            {'hrsh7th/nvim-cmp'},
+            {'hrsh7th/cmp-nvim-lsp'},
+            {'L3MON4D3/LuaSnip'},
+        }
+    },
+    'simrat39/rust-tools.nvim',
+}
