+++ nvim/lua/dmacario/lazy/lsp.lua
@@ -1,6 +1,6 @@
 -- Setup for LSP (with mason + lspconfig) and autocomplete (with cmp)
 local lsp_zero = require('lsp-zero')
-
+lsp_zero.extend_lspconfig()
 lsp_zero.preset('recommended')
