+++ nvim/lua/dmacario/lazy/markdownpreview.lua
@@ -1,0 +1,14 @@
+-- Remaps for markdown preview
+return{
+    "iamcco/markdown-preview.nvim",
+    build = "cd app && npm install",
+    cmd = { "MarkdownPreviewToggle", "MarkdownPreview", "MarkdownPreviewStop" },
+    init = function()
+        vim.g.mkdp_filetypes = { "markdown" }
+    end,
+    ft = { "markdown" },
+    config = function()
+        vim.keymap.set("n", "<leader>mp", vim.cmd.MarkdownPreview)
+        print("Here")
+    end
+}
