+++ nvim/lua/dmacario/lazy/nerdtree.lua
@@ -1,0 +1,7 @@
+return{
+    'preservim/nerdtree',
+    config = function()
+        vim.keymap.set('n', '<leader>o', vim.cmd.NERDTreeToggle)
+        vim.g.NerdTreeShowHidden = true
+    end
+}
