+++ nvim/lua/dmacario/lazy/trouble.lua
@@ -1,0 +1,10 @@
+return{
+    'folke/trouble.nvim',
+    config = function()
+        local signs = { Error = " ", Warning = " ", Hint = " ", Information = " " }
+        for type, icon in pairs(signs) do
+            local hl = "DiagnosticSign" .. type
+            vim.fn.sign_define(hl, { text = icon, texthl = hl, numhl = "" })
+        end
+    end
+}
