+++ nvim/lua/dmacario/lazy_init.lua
@@ -1,0 +1,17 @@
+-- Lazy.nvim setup
+
+local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
+if not vim.loop.fs_stat(lazypath) then
+  vim.fn.system({
+    "git",
+    "clone",
+    "--filter=blob:none",
+    "https://github.com/folke/lazy.nvim.git",
+    "--branch=stable", -- latest stable release
+    lazypath,
+  })
+end
+vim.opt.rtp:prepend(lazypath)
+
+require("lazy").setup({ spec = "dmacario.lazy", change_detection = { notify = false} })
+
